pyproject.toml
[project]
name = "nse-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "groq>=1.1.2",
    "loguru>=0.7.3",
    "logzero>=1.7.0",
    "mcp>=1.27.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "psycopg2>=2.9.11",
    "pyotp>=2.9.0",
    "smartapi-python>=1.5.5",
    "websocket>=0.2.1",
]
